% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartificate-package.R
\docType{package}
\name{quartificate-package}
\alias{quartificate}
\alias{quartificate-package}
\title{quartificate: Transform Google Docs into Quarto Books}
\description{
Automate the Transformation of a Google Document into a Quarto Book source.
}
\author{
\strong{Maintainer}: Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

}
\keyword{internal}
