% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartificate.R
\name{quartificate}
\alias{quartificate}
\title{Transform Google Doc to Quarto book}
\usage{
quartificate(gdoc_id, path, render = FALSE)
}
\arguments{
\item{gdoc_id}{ID of the Google Document.}

\item{path}{A path, where to create the Quarto book. It will be created if needed.}

\item{render}{Logical. Whether to render the resulting Quarto book.}
}
\value{
Nothing.
}
\description{
Transform Google Doc to Quarto book
}
\examples{
id <- googledrive::drive_find(
    q = "name contains 'rOpenSci Champions Program Mentor Guidelines v1.0'"
  )$id
 quarto_dir <- withr::local_tempdir()
 quartificate::quartificate(id, quarto_dir, render = TRUE)
}
